Tighten up the check to determine whether a file is an archive type
Previously we were classing any file that conformed to an archive type as an archive, but that means that files such as .ipa and .docx, which are zip files under the hood, were getting classed as archives and unpicked in Zippity. We now just check whether a file's final UTI is in the list of archive types that Zippity supports.